Rework color storage to be a BIT less insane
The code is still nuts and I hate my past self, but at least the wire
color shows in the failure cases now.

Also refactored some nearby stuff just because I was going to go a
different route with this and touched them in the process. Didn't put
the old code style back because it's awful anyway.